apps/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pytest>=7.0.0
httpx>=0.25.0
